Fix auto layout issues in prototype size calculation by first setting bounds to a non zero size and then calling layoutIfNeeded.  Note that this needs to be done to allow the nested collection view to initialize and pull data from it's delegate, but only needs to be done once, and could be done in something like a lazy initializer.
Also note that while this fixes the auto layout conflict errors, it still does not fix the width of the nested cells, which are supposed to be full width (and are declared as such from the nested collections sizeForItemAt delegate method).  This is to be addressed in the following commits.