Fix issue with nested collection ignoring delegate's sizeForItemAt and using narrower cell widths.
There is a frustrating amount of conflict in the implementation of UICollectionView and the flow layout.  It has been possible to use self sizing cells via constraints, but this runs afoul of the flow layout's memory of where cells are, I have see cases where tapping on the top right of the first cell triggers didSelect on later cells, based presumably, on the flow layout's initial usage of estimated size.  Also, as in this sample, constraint based cell layout wreaks havoc with insertion and deletion animations, all the cells seem to collapse and then snap back to the final layout.  The only way through seemed to be using the sizeThatFits delegate method.  Even with an implementation that returned the desired size though, the collection view still used smaller cells, laying out multiple in one row.  I traced this back to being caused by the collection view overriding the final cell size with values from layout attributes.  It turns out that this in turn, is derived from having an implementation of systemSizeFitting... on the collection view, that is required by the sizeThatFits approach.  Evidently, it over-writes the delegate value with this.

The final cause was that I was adjusting the content size back to the width of the contents in the delegate and not doing all of this in the sizaeThatFits method.  This seems like a bug in the collection view implementation, if you provide a requested size, it should respect that.